requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
flask_cors==4.0.1
huggingface_hub==0.24.6
datasets==2.21.0
pydub==0.25.1
soundfile==0.12.1
psycopg2==2.9.9